Trabalho 6/sistema-carga-ia/generator/requirements.txt
pika==1.2.0
Pillow==9.4.0
python-dotenv==0.21.0
